+++ tests/testthat/myfuns.R
@@ -34,7 +34,7 @@ add <- function(a, b) {
 
 run_with_progress <- function(n, wait) {
   for (i in seq_len(n)) {
-    rrq::rrq_progress_update(sprintf("iteration %d", i))
+    rrq::rrq_task_progress_update(sprintf("iteration %d", i))
     Sys.sleep(wait)
   }
   n
@@ -42,7 +42,7 @@ run_with_progress <- function(n, wait) {
 
 
 run_with_progress_interactive <- function(path, poll = 0.01) {
-  rrq::rrq_progress_update("Waiting for file")
+  rrq::rrq_task_progress_update("Waiting for file")
   while (!file.exists(path)) {
     Sys.sleep(poll)
   }
@@ -51,8 +51,8 @@ run_with_progress_interactive <- function(path, poll = 0.01) {
     if (identical(contents, "STOP")) {
       break
     }
-    rrq::rrq_progress_update(sprintf("Got contents '%s'", contents))
+    rrq::rrq_task_progress_update(sprintf("Got contents '%s'", contents))
     Sys.sleep(poll)
   }
-  rrq::rrq_progress_update("Finishing")
+  rrq::rrq_task_progress_update("Finishing")
 }